+++ else/README.md
@@ -1,0 +1,89 @@
+# noise-spotting-tool
+周期的な模様に混ざるノイズ成分のみ抽出する画像処理．
+なお，ノイズ成分をアーティファクトと呼ぶことにする（こちらも正しい専門用語ではないが）．
+幾つかの手法を試して，比較する．
+
+# 1. 課題概要
+## 1.1. Situation:
+入力画像には、縦線（ボーダー）とアーティファクトが混在しています。
+## 1.2. Complication:
+これらのアーティファクトは、画像の品質や解釈を悪化させる可能性があります。
+アーティファクトが残ると、画像の信頼性が低下し、正確な情報が得られなくなる可能性があります。
+## 1.3. Resolution:
+エッジ検出と除去、アーティファクトの特徴抽出、アーティファクトのみを抽出するといった画像処理手法を使用して、アーティファクトを抽出します。
+最適化を行い、抽出されたアーティファクトの品質を向上させます。
+
+
+# 2. 問題解決法
+# 2.1. エッジ検出と除去:
+
+技術要素:
+
+- SobelフィルタやCannyエッジ検出器などのエッジ検出アルゴリズム
+- 2次微分を使用したラプラシアンフィルタ
+- エッジ検出後に、エッジを除去するフィルタや手法
+
+実装フロー:
+
+1. 入力画像をグレースケールに変換する。
+2. エッジ検出アルゴリズムを適用して、画像内の縦線（ボーダー）を検出する。
+3. 検出されたエッジをマスクとして使用し、元の画像からエッジを除去する。
+4. 残された部分がアーティファクトであるため、これを出力として抽出する。
+
+
+# 2.2. アーティファクトの特徴抽出:
+
+技術要素:
+
+- 色の変化やテクスチャの特徴を抽出する手法（例: ヒストグラム平滑化、テクスチャ解析）
+- 画像の局所的な特徴量を抽出する手法（例: SIFT、SURF、HOG）
+
+実装フロー:
+
+1. 入力画像をグレースケールに変換する。
+2. 画像の局所的な特徴量を抽出する。
+3. 特徴量を使用して、縦線（ボーダー）以外の不要な部分を識別し、アーティファクトの特徴を抽出する。
+4. 特徴を使用して、アーティファクトを出力として抽出する。
+
+
+# 2.3. アーティファクトのみを抽出:
+
+技術要素:
+
+- セグメンテーション手法（例: 閾値処理、領域成長法）
+- バイナリ化やマスキングを用いた画像処理手法
+
+実装フロー:
+
+1. 入力画像をグレースケールに変換する。
+2. アーティファクトの特徴を使用して、縦線（ボーダー）以外の不要な部分を識別する。
+3. 識別された不要な部分を除去し、アーティファクトのみを残す。
+4. アーティファクトを出力として抽出する。
+
+
+# 2.4. 最適化:
+
+技術要素:
+
+- ハイパーパラメータの調整（例: フィルタのサイズ、閾値の設定）
+- アルゴリズムの最適化（例: オプティマイザの選択、反復回数の調整）
+
+実装フロー:
+
+1. 既存の手法やアルゴリズムのパラメータを調整し、最適な結果を得る。
+2. パラメータの組み合わせを評価し、最良の結果を与える組み合わせを特定する。
+3. 最適なパラメータを使用して、アーティファクトを抽出する。
+
+
+# 2.5. 評価:
+
+技術要素:
+
+- PSNR（ピーク信号対雑音比）、SSIM（構造的類似度指標）などの品質評価指標
+- 人間の視覚システムによる視覚評価
+
+実装フロー:
+
+1. 抽出されたアーティファクトの品質を評価するために、適切な評価指標を使用する。
+2. 品質評価指標を使用して、アーティファクトの品質を定量化し、必要に応じて改善する。
+
