+++ else/imgs_edge_removal/output_binary_image_50_80.jpg
