+++ else/imgs_edge_removal/output_image_100_120.jpg
