+++ else/imgs_edge_removal/output_image_50_80.jpg
