+++ else/imgs_edge_removal/output_image_80_100.jpg
